move make_value_key to hashing.py
